Word2Vec_SVM/requirements.txt
pandas
numpy
scikit-learn
gensim
matplotlib
umap-learn
openpyxl

